refactor: extract UI logic into useGameUI composable
